+++ bottom-bar/src/main/java/com/roughike/bottombar/BottomBar.java
@@ -597,12 +597,15 @@ public class BottomBar extends LinearLayout implements View.OnClickListener, Vie
     private void updateSelectedTab(int newPosition) {
         int newTabId = getTabAtPosition(newPosition).getId();
 
-        if (newPosition != currentTabPosition && onTabSelectListener != null) {
-            currentTabPosition = newPosition;
-            onTabSelectListener.onTabSelected(newTabId);
+        if (newPosition != currentTabPosition) {
+            if (onTabSelectListener != null) {
+                onTabSelectListener.onTabSelected(newTabId);
+            }
         } else if (onTabReselectListener != null && !ignoreTabReselectionListener) {
             onTabReselectListener.onTabReSelected(newTabId);
         }
+
+        currentTabPosition = newPosition;
 
         if (ignoreTabReselectionListener) {
             ignoreTabReselectionListener = false;